refactor!: separate parsing from validation
