+++ src/definitions/structure.rs
@@ -1,16 +1,12 @@
 //! Parsing and validation of struct definitions.
-use slang_solidity::cst::{Cursor, Query, QueryMatch, TextRange};
+use slang_solidity::cst::TextRange;
 
 use crate::{
-    error::Result,
-    lint::{Diagnostic, ItemDiagnostics, ItemType},
+    lint::{check_params, Diagnostic, ItemDiagnostics},
     natspec::NatSpec,
 };
 
-use super::{
-    capture, check_params, extract_comment, extract_parent_name, Definition, Identifier, Parent,
-    Validate, ValidationOptions,
-};
+use super::{Identifier, ItemType, Parent, SourceItem, Validate, ValidationOptions};
 
 /// A struct definition
 #[derive(Debug, Clone, bon::Builder)]
@@ -33,7 +29,11 @@ pub struct StructDefinition {
     pub natspec: Option<NatSpec>,
 }
 
-impl Validate for StructDefinition {
+impl SourceItem for StructDefinition {
+    fn item_type() -> ItemType {
+        ItemType::Struct
+    }
+
     fn parent(&self) -> Option<Parent> {
         self.parent.clone()
     }
@@ -45,49 +45,20 @@ impl Validate for StructDefinition {
     fn span(&self) -> TextRange {
         self.span.clone()
     }
+}
 
-    fn query() -> Query {
-        Query::parse(
-            "@struct [StructDefinition
-            @struct_name name:[Identifier]
-            @struct_members members:[StructMembers]
-        ]",
-        )
-        .expect("query should compile")
-    }
-
-    fn extract(m: QueryMatch) -> Result<Definition> {
-        let structure = capture(&m, "struct")?;
-        let name = capture(&m, "struct_name")?;
-        let members = capture(&m, "struct_members")?;
-
-        let span = structure.text_range();
-        let name = name.node().unparse().trim().to_string();
-        let members = extract_struct_members(&members)?;
-        let natspec = extract_comment(&structure.clone(), &[])?;
-        let parent = extract_parent_name(structure);
-
-        Ok(StructDefinition {
-            parent,
-            name,
-            span,
-            members,
-            natspec,
-        }
-        .into())
-    }
-
+impl Validate for StructDefinition {
     fn validate(&self, options: &ValidationOptions) -> ItemDiagnostics {
         let mut out = ItemDiagnostics {
             parent: self.parent(),
-            item_type: ItemType::Struct,
+            item_type: Self::item_type(),
             name: self.name(),
             span: self.span(),
             diags: vec![],
         };
         // raise error if no NatSpec is available
         let Some(natspec) = &self.natspec else {
-            if !options.struct_params && !options.enforce.contains(&ItemType::Struct) {
+            if !options.struct_params && !options.enforce.contains(&Self::item_type()) {
                 return out;
             }
             out.diags.push(Diagnostic {
@@ -104,31 +75,13 @@ impl Validate for StructDefinition {
     }
 }
 
-/// Extract the identifiers for each of a struct's members
-fn extract_struct_members(cursor: &Cursor) -> Result<Vec<Identifier>> {
-    let cursor = cursor.spawn();
-    let mut out = Vec::new();
-    let query = Query::parse(
-        "[StructMember
-        @member_name name:[Identifier]
-    ]",
-    )
-    .expect("query should compile");
-    for m in cursor.query(vec![query]) {
-        let member_name = capture(&m, "member_name")?;
-        out.push(Identifier {
-            name: Some(member_name.node().unparse().trim().to_string()),
-            span: member_name.text_range(),
-        });
-    }
-    Ok(out)
-}
-
 #[cfg(test)]
 mod tests {
     use semver::Version;
     use similar_asserts::assert_eq;
     use slang_solidity::{cst::NonterminalKind, parser::Parser};
+
+    use crate::parser::slang::Extract as _;
 
     use super::*;
