+++ src/validate.rs
@@ -1,0 +1,6 @@
+//! The core `NatSpec` validation logic
+
+use crate::{
+    config::Config,
+    lint::{ItemDiagnostics, ItemType},
+};
