+++ tests/tests-basic-sample.rs
@@ -1,8 +1,8 @@
 use std::path::PathBuf;
 
 use lintspec::{
-    definitions::ValidationOptions,
-    lint::{lint, FileDiagnostics},
+    lint::{lint, FileDiagnostics, ValidationOptions},
+    parser::slang::SlangParser,
     print_reports,
 };
 
@@ -14,7 +14,7 @@ fn generate_output(diags: FileDiagnostics) -> String {
 
 #[test]
 fn test_basic() {
-    let diags = lint(
+    let diags = lint::<SlangParser>(
         "./test-data/BasicSample.sol",
         &ValidationOptions::builder().inheritdoc(false).build(),
         true,
@@ -26,7 +26,7 @@ fn test_basic() {
 
 #[test]
 fn test_inheritdoc() {
-    let diags = lint(
+    let diags = lint::<SlangParser>(
         "./test-data/BasicSample.sol",
         &ValidationOptions::default(),
         true,
@@ -38,7 +38,7 @@ fn test_inheritdoc() {
 
 #[test]
 fn test_constructor() {
-    let diags = lint(
+    let diags = lint::<SlangParser>(
         "./test-data/BasicSample.sol",
         &ValidationOptions::builder()
             .inheritdoc(false)
@@ -53,7 +53,7 @@ fn test_constructor() {
 
 #[test]
 fn test_struct() {
-    let diags = lint(
+    let diags = lint::<SlangParser>(
         "./test-data/BasicSample.sol",
         &ValidationOptions::builder()
             .inheritdoc(false)
@@ -68,7 +68,7 @@ fn test_struct() {
 
 #[test]
 fn test_enum() {
-    let diags = lint(
+    let diags = lint::<SlangParser>(
         "./test-data/BasicSample.sol",
         &ValidationOptions::builder()
             .inheritdoc(false)
