+++ tests/tests-library-sample.rs
@@ -1,8 +1,8 @@
 use std::path::PathBuf;
 
 use lintspec::{
-    definitions::ValidationOptions,
-    lint::{lint, FileDiagnostics},
+    lint::{lint, FileDiagnostics, ValidationOptions},
+    parser::slang::SlangParser,
     print_reports,
 };
 
@@ -14,7 +14,7 @@ fn generate_output(diags: FileDiagnostics) -> String {
 
 #[test]
 fn test_basic() {
-    let diags = lint(
+    let diags = lint::<SlangParser>(
         "./test-data/LibrarySample.sol",
         &ValidationOptions::builder().inheritdoc(false).build(),
         true,
